Add some polish to the CRUD

+++ app/Http/Controllers/SensorsController.php
@@ -2,6 +2,7 @@
 
 namespace App\Http\Controllers;
 
+use App\Models\Room;
 use App\Models\Sensor;
 use App\Models\User;
 use Exception;
@@ -27,6 +28,8 @@ class SensorsController extends Controller
 
     public function index(Request $request)
     {
+        $this->authorize('viewAny', Sensor::class);
+
         /** @var User $user */
         $user = $request->user();
 
@@ -51,7 +54,16 @@ class SensorsController extends Controller
      */
     public function show(Request $request, Sensor $sensor)
     {
-        $sensor->load('room');
+        $this->authorize('view', $sensor);
+
+        $sensor->load(
+            [
+                'room',
+                'latest_reading' => function ($query) {
+                    return $query->latest()->first();
+                },
+            ]
+        );
         return Inertia::render(
             'Sensors/Show',
             [
@@ -64,10 +76,42 @@ class SensorsController extends Controller
      * Show the form for creating a new resource.
      *
      * @param Request $request
+     * @param Sensor $sensor
+     * @return Response|Responsable
+     */
+    public function edit(Request $request, Sensor $sensor)
+    {
+        $this->authorize('update', $sensor);
+
+        $sensor->load(
+            [
+                'room',
+            ]
+        );
+
+        $user = $request->user();
+
+        $rooms = Room::query()->where('user_id', $request->user()->id)->get();
+
+        return Inertia::render(
+            'Sensors/Edit',
+            [
+                'sensor' => $sensor,
+                'rooms' => $rooms,
+            ]
+        );
+    }
+
+    /**
+     * Show the form for creating a new resource.
+     *
+     * @param Request $request
      * @return Response|Responsable
      */
     public function create(Request $request)
     {
+        $this->authorize('create', Sensor::class);
+
         return Inertia::render('Sensors/Create');
     }
 
@@ -80,10 +124,13 @@ class SensorsController extends Controller
      */
     public function store(Request $request)
     {
+        $this->authorize('create', Sensor::class);
+
         $validated = $this->validate(
             $request,
             [
                 'name' => ['required'],
+                'room_id' => ['numeric'],
                 'slug' => [
                     'required',
                     'alpha_dash',
@@ -92,9 +139,21 @@ class SensorsController extends Controller
             ]
         );
 
-
         $sensor = new Sensor($validated);
         $sensor->user_id = $request->user()->id;
+
+        if ($validated['room_id'] == 0) {
+            $validated['room_id'] = null;
+        }
+
+        if ($validated['room_id']) {
+            $room = Room::find($validated['room_id']);
+            $this->authorize('create', $room);
+
+            $sensor->room_id = $room->id;
+        }
+
+
         $sensor->save();
 
         return Redirect::route('sensors.index')
@@ -111,16 +170,29 @@ class SensorsController extends Controller
      */
     public function update(Request $request, Sensor $sensor): Response
     {
+        $this->authorize('update', $sensor);
+
         $validated = $this->validate(
             $request,
             [
                 'name' => ['required'],
+                'slug' => ['alpha_dash'],
+                'room_id' => ['numeric'],
             ]
         );
 
+        if ($validated['room_id'] == 0) {
+            $validated['room_id'] = null;
+        }
+
+        if ($validated['room_id']) {
+            $room = Room::find($validated['room_id']);
+            $this->authorize('create', $room);
+        }
+
         $sensor->update($validated);
 
-        return Redirect::route('sensors.index')
+        return Redirect::route('sensors.show', ['sensor' => $sensor->slug])
             ->with('message', "Sensor updated successfully.");
     }
 
@@ -134,6 +206,8 @@ class SensorsController extends Controller
      */
     public function destroy(Request $request, Sensor $sensor): Response
     {
+        $this->authorize('destroy', $sensor);
+
         $sensor->delete();
 
         return Redirect::route('sensors.index')
